Add base dotfiles
These are the simplest versions of integrated preferences/settings for a
minimalist development environment. Vim keybindings are used throughout,
but there are no plugin managers in use (for Vim, Tmux, or Zsh) yet.

Vim splits and Tmux panes can be resized with the mouse, and there are
reasonably intuitive ways to leverage the system clipboard in Vim and
Tmux via the xclip command.