Adjust settings to improve command-line experience
Bold fonts were disabled for the urxvt terminal to improve terminal
readability when using small/narrow fonts.

The prompt color was changed from green to yellow for both bash and zsh,
so the prompt is easier to recognize when scrolling through the terminal
buffer. This is especially useful when using `git diff` which tends to
contain green text.

Tmux window setting were adjusted to provide more sane behavior.