Update files in `shell/common.d`
Since the glob pattern to match all hidden and non-hidden
files/directories in the current directory is different between bash and
zsh, set the variable `_all_files_glob` to be the appropriate pattern,
so it can be used in common functions and aliases for bash & zsh.

Set the null_glob option for zsh so non-matching glob patterns are
removed from the command-line argument list instead of reporting an
error.

Add functions and aliases for listing files/directories, moving to a
directory that may not exist yet, showing sorted disk usage of files,
ignoring the ".git" directory when using tree, and for shutdown/reboot.