Add the Vundle plugin manager for Vim
The submodule for Vundle was added to `vim/bundle/Vundle.vim`, `.vimrc`
was moved to `vim/vimrc`, and Vundle setup information was added to the
top of `vim/vimrc`.

The `setup.bash` script updates symbolic links and creates the
`~/.plugin_install_dir/vundle` directory to hold the downloads for
plugins added to Vundle.