Add `vim-tmux-navigator` to Vundle and tmux.conf
The `.tmux.conf` file was moved to `tmux/tmux.conf` and the keybindings
were updated to switch panes without using the `^b` prefix. The symbolic
link for `~/.tmux.conf` was updated in `setup.bash`.

The `vim-tmux-navigator` plugin was added to Vundle which provides the
same keybindings for switching between Vim splits, without using the
`^w` prefix.