Add `install_system.bash` script
The `install_system.bash` script is intended to be run immediately after
completing a fresh install of Ubuntu Mini 14.04 (which only installs the
least amount of software to get the system to boot).

The software packages added by this script are mostly light-weight,
command-line development/productivity tools, with the major exceptions
being `chromium-browser`, `firefox`, `vlc`, `audacity`, `gimp`, and
`evince-gtk`.