Add `_find` and `grip-many` functions
The `_find` wrapper function was added to `shell/common`, allowing it to
be used in the `shell/common.d/` scripts. It allows you to find files
based on filename patterns, directory depth, and modification date.

The `grip-many` function was added to `shell/common.d/grip.sh`. It uses
`_find` to build a markdown file containing links to other markdown
files that match depth and time criteria.