Add `findit` script and a bin directory.
Create a `bin` directory and update `setup.bash` to add a symbolic link
for `~/bin`.

Move the contents of the `_find` function from `shell/common` to
`bin/findit` as a separate shell script (allow us to call it with sudo
if necessary).

Update the `grip-many` function in `shell/common.d/grip.sh` to use the
`findit` script.