Update `repo-list` function and user edit gitconfig after install
Allow repo-list to accept a value for find's -maxdepth option, instead
of forcing a depth of 3. Also prevent the `dirname` command from showing
an error if there are no repos found.

When using the `install_system.bash` script to clone the dotfiles repo
and run it's `setup.bash`, prompt the user to update the email
address and name they want to use for git. Then use vim to open the
gitconfig file no matter what they say.