Add `newscript` script, `p` alias, & other minor updates
The `bin/newscript` script will generate a bash script in ~/bin with
execute permission and open it with vim.

The `p` alias is a wrapper to the `ps` command.

The `Tmux` function no longer creates a session name based on the
current directory if no session name is provided.

The `bin/mystats` script has more useful output with cleaner
organization.