Add `bin/newpy` script
The `newpy` script creates a new directory in `~/py/components`.
Skeleton files `setup.bash` and `requirements.txt` and copied to the new
directory, and vim is started.

The dotfiles `setup.bash` script was updated to create and link the `py`
directory.

The `common` script also sets the PYTHONPATH variable to include
`~/py/components`.