Add `prompt-verbose` function to quickly change prompt
For most instances, the default minimal prompt is sufficient, but it is
useful to have a more verbose prompt (that shows username, hostname,
absolute path, and current time) sometimes.

An example where it's useful to use the more verbose prompt is when you
have multiple tmux splits for the current window, with one of them being
an SSH session to another host.