Update the `helpme` function to produce more useful output
If called with no arguments (or with "helpme" as an argument) describe
how to invoke `helpme` and how to navigate `less`.

If the name of a shell function is passed to `helpme`, print whether
`which` or `type -a` was used to get information (different behavior for
those commands if your shell is Bash or Zsh).

If the name of an alias is passed in, just add the output of
command -v` to the output file.

Finally, show that the `shgrep` script is what is being used to get the
grep output at the end of the output file (which shows where in these
dotfiles the particular the argument to `helpme` is used.